fix: Revert visit_faces to use i32 to match Octree and NeighborGrid types
The visit_faces function and FaceInfo struct were recently changed to use
u8 types, but this caused type mismatches with the rest of the codebase
which uses i32 for voxel materials:

- Octree uses Cube<i32>
- NeighborGrid uses Cube<i32> internally
- generate_face_mesh callers pass Cube<i32>

Changes:
- Reverted visit_faces signature to accept Cube<i32> and [i32; 4]
- Reverted visit_faces_old to match (deprecated function)
- Changed FaceInfo.material_id from u8 to i32
- Reverted generate_face_mesh to use i32 types
- Removed unused max_depth parameter in visit_faces call
- Kept cast to u8 in add_textured_face for MeshBuilder API

This fixes build errors in WASM compilation where type mismatches
prevented the project from compiling.

+++ crates/cube/src/mesh/generator.rs
@@ -133,15 +133,15 @@ impl MeshBuilder for DefaultMeshBuilder {
 /// * `border_materials` - Array of 4 material IDs for border voxels at each Y layer [y0, y1, y2, y3]
 /// * `base_depth` - The depth at which voxels are 1 unit in size (for UV scaling)
 pub fn generate_face_mesh<B, F>(
-    root: &Cube<u8>,
+    root: &Cube<i32>,
     builder: &mut B,
     color_fn: F,
     _max_depth: u32,
-    border_materials: [u8; 4],
+    border_materials: [i32; 4],
     base_depth: u32,
 ) where
     B: MeshBuilder,
-    F: Fn(u8) -> [f32; 3] + Copy,
+    F: Fn(i32) -> [f32; 3] + Copy,
 {
     // Use visit_faces to iterate through all visible faces
     visit_faces(
